holy fuck it actually goes parallel now
the key thing to do is to force work upon forcing (-> WHNF) a value. here the
new Thought structure has a strict constructor so work is done when it's forced.
which means that the `par` in Main now actually sparks it. in the previous
version, we had a Builder and so if you forced it practically nothing happened.