handwrite parser
the BS construction / teardown cost was unavoidable, so I've handwritten a
parser.

alloc ratio between knap.entry and main.work.rep changes from 84 : 13 (out of
97) ~ 6.46 odds to 88 : 10 (out of 98) ~ 8.80 odds

measurements of course fluctuate, but the ratio improvement is really
noticeable.

this doesn't exactly carry over to improved latency or noticeably less
allocation.

but this is good for the GC, so the improvements are probably latent