apply @effectfully's suggestions (0.33 s)
regression (+ 0.02 s). regression cannot eliminate chance element.

in lines 104-106 in Knap.hs, it does not matter whether lines 105 and 106 have
bangs added to the 0 and w parameters, respectively. that is, if the only bang
added is in line 104 (for the wildcard parameter), the runtime is still 0.33 s.

local benchmarks indicate that allocation is down (~ 15 M -> 11 M), but latency
(wall clock time) is mostly unchanged.